requirements.txt
discord.py
discord.py[voice]
youtube-dl